refactor(WelcomeTemplate): create base, electron and web versions of WelcomeTemplate component
